Implementing convolutional neural networks
This code defines a class ConvNetwork that extends the Network class and adds convolutional and pooling layers to implement a convolutional neural network (CNN) for image classification tasks. The CNN architecture consists of a single convolutional layer with 20 filters of size 5x5, a max pooling layer with pool size of 2, a fully connected layer, and an output layer with softmax activation function. The weights and biases of the fully connected layer are randomly initialized using Gaussian distribution.

The feedforward method performs the forward pass of the CNN by applying the convolutional layer with ReLU activation function, the max pooling layer, the fully connected layer with ReLU activation function, and the output layer with softmax activation function. The input is reshaped to have a shape of (28, 28, 1) to fit the input shape of the network.

The evaluate method computes the cost and accuracy on the provided data. For each input in the data, the method computes the predicted output using feedforward and compares it with the true label to calculate the accuracy. It also computes the total cost by applying the loss function on the predicted output and the true label. Finally, the method returns the average cost and accuracy on the data.
However, the loss argument in the constructor of ConvNetwork class specifies the loss function to be used for training the network. The default loss function is CrossEntropyLoss. Also, the init argument specifies the initialization method for the weights and biases of the fully connected layer. The default initialization method is 'standard'.