Create Brownian Motion (animation)
This code generates a Brownian motion animation using the matplotlib.animation module. The brownian_motion function takes in the number of discrete steps (N), the number of continuous time steps (T), and the variance of the increments (h) and returns an array of the Brownian motion and an array of the random increments.